completed the create route functionality, from front end to back end

+++ server/db/migrations/20231219215632_cart.js
@@ -7,7 +7,7 @@ exports.up = function (knex) {
     table.increments();
     table.string('name');
     table.string('image_url');
-    table.integer('price');
+    table.float('price');
     table.integer('quantity');
   });
 };
